refactor(fund-dashboard): flatten dashboard modules into app root
Move the fund dashboard providers, services, hooks, storage, and UI components out of app/features/fund-dashboard and into the main app structure.

Consolidate shared dashboard types into app/types, update page and test imports, add local AI tooling ignore rules, add the skills lock file, and remove the old CNAME entry.
