+++ .gitignore
@@ -83,9 +83,14 @@ fabric.properties
 .npm-cache/
 .pnpm-store/
 .sisyphus/
+.claude/
+.codex/
+.agents/
 .vercel/
 *.tsbuildinfo
 .impeccable.md
+openspec/
+skills/
 coverage/
 playwright-report/
 test-results/
