--- CNAME
@@ -1,1 +1,0 @@
-fund-baby.ningzhengsheng.cn
