+++ app/api/providers/fund-parsers.ts
@@ -181,3 +181,4 @@ export function parseTrendSummary(trend: TrendPoint[]) {
         : null,
   };
 }
+
