+++ app/components/add-fund-panel.tsx
@@ -209,3 +209,4 @@ export function AddFundPanel({
     </div>
   );
 }
+
