+++ app/components/dashboard-filter-bar.tsx
@@ -111,3 +111,4 @@ export function DashboardFilterBar({
     </div>
   );
 }
+
