+++ app/components/dashboard-management-modals.tsx
@@ -1,0 +1,6 @@
+'use client';
+
+export { AddResultModal } from '@/app/components/add-result-modal';
+export { ConfirmModal } from '@/app/components/confirm-modal';
+export { SuccessModal } from '@/app/components/success-modal';
+
