+++ app/components/dashboard-trade-modals.tsx
@@ -1,0 +1,12 @@
+'use client';
+
+export type {
+  BuyTradeConfirmData,
+  SellTradeConfirmData,
+  TradeConfirmData,
+} from '@/app/components/trade-modal';
+export { HoldingActionModal } from '@/app/components/holding-action-modal';
+export { HoldingEditModal } from '@/app/components/holding-edit-modal';
+export { TopStocksModal } from '@/app/components/top-stocks-modal';
+export { TradeModal } from '@/app/components/trade-modal';
+
