+++ app/components/group-summary.tsx
@@ -346,3 +346,4 @@ export function GroupSummary({
     </div>
   );
 }
+
