+++ app/components/holding-action-modal.tsx
@@ -97,3 +97,4 @@ export function HoldingActionModal({
     </motion.div>
   );
 }
+
