--- app/features/fund-dashboard/components/dashboard-management-modals.tsx
@@ -1,5 +1,0 @@
-'use client';
-
-export { AddResultModal } from '@/app/features/fund-dashboard/components/add-result-modal';
-export { ConfirmModal } from '@/app/features/fund-dashboard/components/confirm-modal';
-export { SuccessModal } from '@/app/features/fund-dashboard/components/success-modal';
