--- app/features/fund-dashboard/components/dashboard-trade-modals.tsx
@@ -1,11 +1,0 @@
-'use client';
-
-export type {
-  BuyTradeConfirmData,
-  SellTradeConfirmData,
-  TradeConfirmData,
-} from '@/app/features/fund-dashboard/components/trade-modal';
-export { HoldingActionModal } from '@/app/features/fund-dashboard/components/holding-action-modal';
-export { HoldingEditModal } from '@/app/features/fund-dashboard/components/holding-edit-modal';
-export { TopStocksModal } from '@/app/features/fund-dashboard/components/top-stocks-modal';
-export { TradeModal } from '@/app/features/fund-dashboard/components/trade-modal';
