--- app/features/fund-dashboard/fund-dashboard-page.tsx
@@ -1,5 +1,0 @@
-import FundDashboardClient from '@/app/features/fund-dashboard/fund-dashboard-client';
-
-export default function FundDashboardPage() {
-  return <FundDashboardClient />;
-}
