+++ app/hooks/use-fund-search.ts
@@ -102,3 +102,4 @@ export function useFundSearch({
     toggleSelectFund,
   };
 }
+
