+++ app/page.tsx
@@ -1,5 +1,6 @@
-import FundDashboardPage from '@/app/features/fund-dashboard/fund-dashboard-page';
+import FundDashboardClient from '@/app/fund-dashboard-client';
 
 export default function Page() {
-  return <FundDashboardPage />;
+  return <FundDashboardClient />;
 }
+
